support archetype
