--- connectors/rocketmq-connect-clickhouse/src/main/java/org/apache/rocketmq/connect/clickhouse/config/ClickHouseSinkConfig.java
@@ -1,30 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.rocketmq.connect.clickhouse.config;
-
-import java.util.HashSet;
-import java.util.Set;
-
-public class ClickHouseSinkConfig extends ClickHouseBaseConfig {
-    public static final Set<String> SINK_REQUEST_CONFIG = new HashSet<String>() {
-        {
-            add(ClickHouseConstants.CLICKHOUSE_HOST);
-            add(ClickHouseConstants.CLICKHOUSE_PORT);
-        }
-    };
-}
