+++ connectors/rocketmq-connect-connectors-archetype/src/main/resources/archetype-resources/src/test/java/source/__dbNameToCamel__SourceTaskTest.java
@@ -1,0 +1,28 @@
+#set( $symbol_pound = '#' )
+#set( $symbol_dollar = '$' )
+#set( $symbol_escape = '\' )
+package ${package}.source;
+
+import io.openmessaging.KeyValue;
+import io.openmessaging.connector.api.data.ConnectRecord;
+import io.openmessaging.internal.DefaultKeyValue;
+import java.util.List;
+import junit.framework.TestCase;
+import ${package}.config.${dbNameToCamel}Constants;
+
+import static java.lang.Thread.sleep;
+
+public class ${dbNameToCamel}SourceTaskTest {
+
+//    private static final String host = "127.0.0.1";
+//    private static final String port = "8123";
+//    private static final String db = "default";
+//    private static final String username = "default";
+//    private static final String password = "123456";
+//
+//    public void testPoll() {
+//    }
+//
+//    public void testStart() throws InterruptedException {
+//    }
+}
