+++ connectors/rocketmq-connect-connectors-archetype/src/test/resources/projects/basic/archetype.properties
@@ -1,0 +1,11 @@
+#Mon Aug 14 22:31:14 CST 2023
+package=it.pkg
+version=0.1-SNAPSHOT
+groupId=archetype.it
+artifactId=basic
+databaseName=test
+dbNameToUpperCase=TEST
+dbNameToCamel=Test
+dbNameToLowerCase=test
+archetype.filteredExtensions=java,sql,yml,xml,properties,factories,ftl,md
+
