+++ connectors/rocketmq-connect-connectors-archetype3/README.md
@@ -1,0 +1,1 @@
+##### How to use archetype
